requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.0
python-dotenv==1.0.0
pandas
numpy
scikit-learn==1.3.2
joblib==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
pydantic==2.5.2
pydantic[email]==2.5.2
pyjwt==2.8.1
xgboost==2.0.0
lightgbm==4.1.0
torch==2.1.1
tensorflow==2.14.0
statsmodels==0.14.0
